refactor: split dict connection I/O and separate dict + user_history tests
- Extract connection I/O methods into dict/connection_io.rs (284 lines)
- Reduce dict/connection.rs from 677 to 111 lines (struct + accessors)
- Move dict tests to dict/tests/ (connection: 292, trie_dict: 282 lines)
- Reduce trie_dict.rs from 476 to 190 lines
- Move user_history tests to user_history/tests.rs (256 lines)
- Reduce user_history/mod.rs from 594 to 340 lines
