+++ engine/src/dict/mod.rs
@@ -1,6 +1,9 @@
 pub mod connection;
+mod connection_io;
 mod entry;
 pub mod source;
+#[cfg(test)]
+mod tests;
 mod trie_dict;
 
 pub use entry::DictEntry;
