+++ engine/src/dict/tests/mod.rs
@@ -1,0 +1,2 @@
+mod connection;
+mod trie_dict;
